feat(ClickableMap): allow user to enter address
We now allow users to enter address on the new request and volunteer sign up pages. On the
volunteer sign up page, the address also populates the separate address fields for the volunteer.

Closes #29

fix(NewRequestPage): fix the .length error and styling

Fix an error when the user doesn't have the diplayName filled out. Also default displayName
to `firstName lastName` when it's not set. Lastly, fix the styling on small screens.

Closes #85

feat(RequestSuccessfulPage): add link to the newly created request