Address lookup (#96)
* fix(SearchPage): fix the responsive styling of the results
* fix(ClickableMap): switch to calling Google Maps API directly

Previously, we were using the react-geocode library which was a server side library and did not
support domain-based key restrictions. Now using the API directly because it's simple enough and
will allow us to support key restrictions.

Closes #94

* chore(*): upgrade a few of the older packages.
* refactor(AccountMenu): switch to Link component to allow opening in new window
* feat(ClickableMap): allow user to enter address

We now allow users to enter address on the new request and volunteer sign up pages. On the
volunteer sign up page, the address also populates the separate address fields for the volunteer.

Closes #29

* fix(NewRequestPage): fix the .length error and styling

Fix an error when the user doesn't have the diplayName filled out. Also default displayName
to `firstName lastName` when it's not set. Lastly, fix the styling on small screens.

Closes #85

* feat(RequestSuccessfulPage): add link to the newly created request
* feat(*): hook up sentry.io. (#95)
* fix(Discussion): remove the incorrect limit(1) from private comments.
* chore(*): update version
* Check version before creating new release (#100)
* Only run release logic for production branch

* refactor(ClickableMap): switch to async/await as suggested
* refactor(useNewRequest): copy the requestLocation object before deleting the unneeded property
* refactor(utils/geo): log to Sentry when API is not found

Co-authored-by: Scott Prue <prescottprue@users.noreply.github.com>